+++ apps/accounts/api/v1/urls.py
@@ -4,7 +4,6 @@ from django.urls import include, path
 
 from .views import (
     AuthViewSet,
-    OrderHistoryViewSet,
     SMSAuthViewSet,
     UserDetailView,
     UserProfileViewSet,
@@ -16,7 +15,6 @@ router = DefaultRouter()
 router.register(r"profile", UserProfileViewSet, basename="profile")
 router.register(r"auth", AuthViewSet, basename="auth")
 router.register(r"sms-auth", SMSAuthViewSet, basename="sms-auth")
-router.register(r"orders", OrderHistoryViewSet, basename="orders")
 
 urlpatterns = [
     path("", include(router.urls)),
