+++ apps/orders/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Order
+
+admin.site.register(Order)
