+++ apps/orders/api/v1/urls.py
@@ -6,8 +6,8 @@ from . import views
 
 router = routers.DefaultRouter()
 
-router.register(r"records", views.RecordViewSet)
-router.register(r"styles", views.StyleViewSet)
+router.register(r"orders", views.OrderViewSet)
+
 urlpatterns = [
     path("", include(router.urls)),
 ]
