+++ apps/records/api/v1/views.py
@@ -3,10 +3,10 @@ from rest_framework.response import Response
 
 from django.shortcuts import get_object_or_404
 
-from records.models import Order, Record, Style
+from records.models import Record, Style
 
 from .filters import RecordFilter
-from .serializers import OrderSerializer, RecordSerializer, StyleSerializer
+from .serializers import RecordSerializer, StyleSerializer
 
 
 class RecordViewSet(viewsets.ReadOnlyModelViewSet):
@@ -43,28 +43,3 @@ class StyleViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Style.objects.all()
     serializer_class = StyleSerializer
     pagination_class = None
-
-
-class OrderViewSet(viewsets.ModelViewSet):
-    queryset = Order.objects.all()
-    serializer_class = OrderSerializer
-    http_method_names = ["get", "post"]
-    permission_classes = []
-
-    def get_queryset(self):
-        if self.request.user.is_authenticated:
-            return Order.objects.filter(user=self.request.user)
-        return Order.objects.none()
-
-    def get_object(self):
-        return get_object_or_404(self.get_queryset(), pk=self.kwargs.get("pk"))
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        try:
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        except exceptions.ValidationError as err:
-            return Response({"error": str(err)}, status=status.HTTP_400_BAD_REQUEST)
